requirements.txt
anthropic>=0.42.0
openai>=1.60.0
PyGithub>=2.5.0
langfuse>=2.56.0
pyyaml>=6.0.2
flask>=3.1.0
gunicorn>=23.0.0
kubernetes>=31.0.0
# OpenTelemetry
opentelemetry-api>=1.29.0
opentelemetry-sdk>=1.29.0
opentelemetry-exporter-otlp-proto-grpc>=1.29.0
opentelemetry-exporter-prometheus>=0.50b0
prometheus-client>=0.21.0
requests>=2.32.0
